set budget details

+++ app/(func)/budgetDetail.tsx
@@ -1,8 +1,9 @@
-import { Text, View, TouchableOpacity, TextInput } from "react-native";
+import { Text, View, TouchableOpacity, TextInput, Alert } from "react-native";
 import React, { useEffect, useState } from "react";
 import { useLocalSearchParams, useRouter } from "expo-router";
 import { useTheme } from "../../contexts/ThemeContext";
 import { deleteBudget, updateBudget } from "@/services/budgetService";
+import { Ionicons } from "@expo/vector-icons";
 
 const BudgetDetail = () => {
   const { theme } = useTheme();
@@ -19,33 +20,66 @@ const BudgetDetail = () => {
   }, [amount]);
 
   const handleDelete = async () => {
-    try {
-      await deleteBudget(budgetId as string);
-      router.back();
-    } catch (error) {
-      console.error("删除预算失败:", error);
-    }
+    Alert.alert(
+      "确认删除",
+      "您确定要删除这个预算吗？",
+      [
+        {
+          text: "取消",
+          onPress: () => console.log("删除操作已取消"),
+          style: "cancel",
+        },
+        {
+          text: "确定",
+          onPress: async () => {
+            try {
+              await deleteBudget(budgetId as string);
+              router.back();
+            } catch (error) {
+              console.error("删除预算失败:", error);
+            }
+          },
+        },
+      ],
+      { cancelable: false },
+    );
   };
 
   const handleUpdate = () => {
     if (isEditing && newAmount !== String(amount)) {
-      // Construct the updated data object
-      const updatedData = {
-        category: category as string,
-        amount: parseFloat(newAmount), // Update the amount
-      };
-      console.log("更新预算，ID:", budgetId);
+      Alert.alert(
+        "确认更新",
+        "您确定要保存这些更改吗？",
+        [
+          {
+            text: "取消",
+            onPress: () => console.log("更新操作已取消"),
+            style: "cancel",
+          },
+          {
+            text: "确定",
+            onPress: () => {
+              const updatedData = {
+                category: category as string,
+                amount: parseFloat(newAmount), // Update the amount
+              };
+              console.log("更新预算，ID:", budgetId);
 
-      // Call the updateBudget function with the budgetId and updatedData
-      updateBudget(budgetId as string, updatedData)
-        .then(() => {
-          // After successful update, save the new amount and switch off editing mode
-          setIsEditing(false);
-          setNewAmount(String(updatedData.amount)); // Update the state with the new amount
-        })
-        .catch((error) => {
-          console.error("更新预算失败:", error);
-        });
+              // Call the updateBudget function with the budgetId and updatedData
+              updateBudget(budgetId as string, updatedData)
+                .then(() => {
+                  // After successful update, save the new amount and switch off editing mode
+                  setIsEditing(false);
+                  setNewAmount(String(updatedData.amount)); // Update the state with the new amount
+                })
+                .catch((error) => {
+                  console.error("更新预算失败:", error);
+                });
+            },
+          },
+        ],
+        { cancelable: false },
+      );
     } else {
       setIsEditing(true); // Switch to edit mode
     }
@@ -106,6 +140,27 @@ const BudgetDetail = () => {
           <Text className="text-lg font-semibold text-white">删除预算</Text>
         </TouchableOpacity>
       </View>
+
+      {/* 如果预算为0，显示 Set Budget 按钮 */}
+      {parseFloat(newAmount) === 0 && (
+        <TouchableOpacity
+          onPress={() => router.push("/(func)/budget")}
+          className={`flex-row items-center p-3 mt-6 rounded-xl shadow-md ${
+            theme === "dark" ? "bg-quaternary" : "bg-white"
+          }`}>
+          <Ionicons
+            name="wallet-outline"
+            size={20}
+            color={theme === "dark" ? "#1477f1" : "#0d6df4"}
+          />
+          <Text
+            className={`ml-4 text-lg font-semibold ${
+              theme === "dark" ? "text-gray-200" : "text-gray-700"
+            }`}>
+            设置预算
+          </Text>
+        </TouchableOpacity>
+      )}
     </View>
   );
 };
